+++ database/migrations/2020_02_29_103129_user_rank.php
@@ -14,22 +14,22 @@ class UserRank extends Migration
     public function up()
     {
         schema::create('user_reviews',function(Blueprint $table){
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->integer('order_restaurant_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('order_restaurant_id')->unsigned();
             $table->text('review_text');
             $table->integer('review_rank');
             $table->boolean('isActive');
             $table->timestamps();
-            
+
              $table->foreign('user_id')->references('id')->on('users')
                      ->onUpdate('cascade')->onDelete('cascade');
-             
+
              $table->foreign('order_restaurant_id')->references('id')->on('order_restaurants')
                      ->onUpdate('cascade')->onDelete('cascade');
 
         });
-        
+
     }
 
     /**
