+++ database/migrations/2020_02_29_115033_app_reviews.php
@@ -14,13 +14,13 @@ class AppReviews extends Migration
     public function up()
     {
        schema::create('app_reviews',function(Blueprint $table){
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->text('review_text');
             $table->integer('review_rank');
             $table->boolean('isActive');
             $table->timestamps();
-            
+
              $table->foreign('user_id')->references('id')->on('users')
                      ->onUpdate('cascade')->onDelete('cascade');
